Intereactive Side menu + personal profile prototype done
