+++ Pow Wow/Pow Wow/ViewControllers/HomeViewController.swift
@@ -23,6 +23,9 @@ class HomeViewController: UIViewController {
     @IBOutlet weak var ConsultantCollectionView: UICollectionView!
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        refView = self
+        definesPresentationContext = true
         // Initialize the SideMenu When it loads
         menu = SideMenuNavigationController(rootViewController: MenuListController())
         // Menu will pop up at left
@@ -72,42 +75,6 @@ class HomeViewController: UIViewController {
     }
     */
 
-}
-
-class MenuListController: UITableViewController {
-    
-    var items = ["Consultant Profile", "Business Profile", "Setting", "Log Out", "About Us"]
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        tableView.backgroundColor = darkColor
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-    }
-    
-    let darkColor = UIColor(red: 33/255.0,
-                            green: 33/255.0,
-                            blue: 33/255.0, alpha: 1)
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return items.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        cell.textLabel?.text = items[indexPath.row]
-        cell.textLabel?.textColor = .white
-        cell.backgroundColor = darkColor
-        return cell
-    }
-    
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let refView = HomeViewController()
-        if (items[indexPath.row] == "Consultant Profile") {
-            refView.performSegue(withIdentifier: "HomeToPersonalProfile", sender: self)
-        }
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-    
 }
 
 // Extending HomeViewController for a picker
